Packed relevant structs used in the Attribute and Modifier components. Also fixed a variable naming error that caused the project to fail to build.
