--- src/components/attributes/Life.hpp
@@ -1,5 +1,0 @@
-#pragma once
-
-#include "components/attributes/Attribute.hpp"
-
-struct Life : Attribute {};
