--- src/components/modifiers/FlatLife.hpp
@@ -1,7 +1,0 @@
-#pragma once
-
-#include "components/modifiers/FlatModifier.hpp"
-
-struct FlatLife : FlatModifier {
-  FlatLife(int v) : FlatModifier(v) { id_ = "FLAT_LIFE"; }
-};
