--- src/components/modifiers/MultiplicativeLife.hpp
@@ -1,8 +1,0 @@
-#pragma once
-
-#include "components/modifiers/MultiplicativeModifier.hpp"
-
-struct MultiplicativeLife : MultiplicativeModifier {
-  MultiplicativeLife() : MultiplicativeModifier() { id_ = "MULTIPLICATIVE_LIFE"; }
-  MultiplicativeLife(int v) : MultiplicativeModifier(v) { id_ = "MULTIPLICATIVE_LIFE"; }
-};
